pyproject.toml
[project]
name = "web_fillers"
version = "3.0.0"
description = "Fillers for filling assignments in the seek portal and replit"
readme = "README.md"
requires-python = ">=3.7"
license = {file = "LICENSE.txt"}

authors = [
  {name = "Livin Nector", email = "livin@study.iitm.ac.in" }
]

dependencies = [
  "selenium","markdown-to-json","pyyaml","bs4", "marko"
]

[project.optional-dependencies]

[project.urls]
[project.scripts]
proq = "proq_filler.cli:main"

[tool.setuptools]

[build-system]
requires = ["setuptools>=43.0.0", "wheel"]
build-backend = "setuptools.build_meta"